fix bug

+++ Business/Service/ServiceTable.cs
@@ -37,8 +37,17 @@ namespace Business.Service
         public Table Delete(int id)
         {
             Table tableDel = _tableRepository.GetOne(t => t.Id == id);
-            _tableRepository.Delete(tableDel);
-            return tableDel;
+            if (tableDel == null)
+            {
+                Console.WriteLine("bele idli table tapilmadi");
+                return null;
+            }
+            else
+            {
+                _tableRepository.Delete(tableDel);
+                return tableDel;
+            }
+           
         }
 
         public List<Table> GetAll()
@@ -63,30 +72,58 @@ namespace Business.Service
 
         public Table GetTable(int id)
         {
+            
             Table getTable = _tableRepository.GetOne(t => t.Id == id);
-            Extension.Print(ConsoleColor.Green, $"Id:{getTable.Id}\nTableNo:{getTable.TableNo}\nMoney Table:{getTable.MoneyTabel} azn\n" +
-                                       $"Reservition:{getTable.Reservition}\n" +
-                                       $"Ugurla Update olundu");
-            return getTable;
+            if (getTable == null)
+            {
+                Console.WriteLine("bele idli table yoxdur");
+                return null;
+            }
+            else
+            {
+                Extension.Print(ConsoleColor.Green, $"Id:{getTable.Id}\nTableNo:{getTable.TableNo}\nMoney Table:{getTable.MoneyTabel} azn\n" +
+                                                  $"Reservition:{getTable.Reservition}\n" +
+                                                  $"Ugurla Update olundu");
+                return getTable;
+            }
+       
         }
 
         public Table UpdateMoneyTable(int id, int money)
         {
             Table upTabelMoney = _tableRepository.GetOne(t => t.Id == id);
-            upTabelMoney.MoneyTabel = money;
-            Extension.Print(ConsoleColor.Green, $"Id:{upTabelMoney.Id}\nTableNo:{upTabelMoney.TableNo}\nMoney Table:{upTabelMoney.MoneyTabel} azn\n" +
-                                         $"Reservition:{upTabelMoney.Reservition}\n" +
-                                         $"Ugurla Update olundu");
-            return upTabelMoney;
+            if(upTabelMoney == null)
+            {
+                Console.WriteLine("bele idli table tapilmadi");
+                return null;
+            }
+            else
+            {
+                upTabelMoney.MoneyTabel = money;
+                Extension.Print(ConsoleColor.Green, $"Id:{upTabelMoney.Id}\nTableNo:{upTabelMoney.TableNo}\nMoney Table:{upTabelMoney.MoneyTabel} azn\n" +
+                                             $"Reservition:{upTabelMoney.Reservition}\n" +
+                                             $"Ugurla Update olundu");
+                return upTabelMoney;
+            }
+        
         }
 
         public Table UpdateTableNo(int id, string tableNo)
         {
             Table upTableNo = _tableRepository.GetOne(t => t.Id == id);
-            upTableNo.TableNo = tableNo;
-            Extension.Print(ConsoleColor.Green, $"Id:{upTableNo.Id}\nTableNo:{upTableNo.TableNo}\nMoney Table:{upTableNo.MoneyTabel} azn\n" +
-                                            $"Reservition:{upTableNo.Reservition}\n" +
-                                            $"Ugurla Update olundu");
+            if(upTableNo == null)
+            {
+                Console.WriteLine("bele idli table tapilmadi");
+                return null;
+            }
+            else
+            {
+                upTableNo.TableNo = tableNo;
+                Extension.Print(ConsoleColor.Green, $"Id:{upTableNo.Id}\nTableNo:{upTableNo.TableNo}\nMoney Table:{upTableNo.MoneyTabel} azn\n" +
+                                                $"Reservition:{upTableNo.Reservition}\n" +
+                                                $"Ugurla Update olundu");
+            }
+        
             return upTableNo;
         }
     }
